resolved merge conflicts with new dev branch
